+++ src/Map/utils.test.ts
@@ -1,7 +1,7 @@
-import { act, renderHook } from "@testing-library/react";
+import { renderHook } from "@testing-library/react";
 import { expect, test } from "vitest";
 
-import { useGet, wrapRef, useOverriddenProperty } from "./utils";
+import { useGet, wrapRef } from "./utils";
 
 test("useGet", () => {
   const obj = { a: 1 };
@@ -59,36 +59,4 @@ test("wrapRef", () => {
   expect(res.a(1, 2)).toBeUndefined();
   expect((res as any).b).toBeUndefined();
   expect(res.c(1, 2)).toBeUndefined();
-});
-
-// test for useOverriddenProperty
-test("overriddenProperty", () => {
-  const { result } = renderHook(() => useOverriddenProperty({ a: 1, b: 2, c: { c1: 3 } }));
-  const [mergedProperty, overrideProperty] = result.current;
-  expect(mergedProperty).toEqual({ a: 1, b: 2, c: { c1: 3 } });
-
-  act(() => {
-    overrideProperty("plugin1", { a: 2 } as any);
-  });
-  expect(result.current[0]).toEqual({ a: 2, b: 2, c: { c1: 3 } });
-
-  act(() => {
-    overrideProperty("plugin2", { b: 3 } as any);
-  });
-  expect(result.current[0]).toEqual({ a: 2, b: 3, c: { c1: 3 } });
-
-  act(() => {
-    overrideProperty("plugin3", { c: { c2: 1 } } as any);
-  });
-  expect(result.current[0]).toEqual({ a: 2, b: 3, c: { c1: 3, c2: 1 } });
-
-  act(() => {
-    overrideProperty("plugin1", undefined as any);
-  });
-  expect(result.current[0]).toEqual({ a: 1, b: 3, c: { c1: 3, c2: 1 } });
-
-  act(() => {
-    overrideProperty("plugin2", undefined as any);
-  });
-  expect(result.current[0]).toEqual({ a: 1, b: 2, c: { c1: 3, c2: 1 } });
 });
