+++ src/Visualizer/coreContext.tsx
@@ -1,11 +1,6 @@
 import { createContext } from "react";
 
-import {
-  LayerEditEvent,
-  LayerSelectionReason,
-  LayerVisibilityEvent,
-  ViewerProperty,
-} from "../engines";
+import { LayerEditEvent, LayerSelectionReason, LayerVisibilityEvent } from "../engines";
 import { ComputedFeature, ComputedLayer } from "../mantle";
 import {
   LayerLoadEvent,
@@ -28,8 +23,8 @@ type CoreContext = {
   };
   selectedComputedFeature?: ComputedFeature | undefined;
   viewport?: Viewport;
-  overriddenViewerProperty?: ViewerProperty;
-  overrideViewerProperty?: (pluginId: string, property: ViewerProperty) => void;
+  // overriddenViewerProperty?: ViewerProperty;
+  // overrideViewerProperty?: (pluginId: string, property: ViewerProperty) => void;
   handleCameraForceHorizontalRollChange?: (enable?: boolean) => void;
   handleInteractionModeChange?: (mode?: InteractionModeType | undefined) => void;
   onSketchPluginFeatureCreate?: (cb: SketchEventCallback) => void;
